shifting perf test
