+++ test/amp_metadata.json
@@ -1,1 +1,1 @@
-{"ampRuntimeVersion":"011905140117570","ampCssUrl":"https://cdn.ampproject.org/rtv/011905140117570/v0.css","canaryPercentage":"0.015","diversions":["001905222334000","031905222334000","021905140117570"]}
+{"ampRuntimeVersion":"011906051812580","ampCssUrl":"https://cdn.ampproject.org/rtv/011906051812580/v0.css","canaryPercentage":"0.015","diversions":["001906111828200","031906111828200","021906051812580"]}
